// lib/types/profile.ts
export interface Profile {
  address: string | null
  avatar_url: string | null
  bio: string | null
  business_name: string | null
  city: string | null
  created_at: string | null
  email: string | null
  full_name: string | null
  id: string
  password: string | null
  pincode: string | null
  rating: number | null
  role: string
  updated_at: string | null
  logo: string | null
  }